# developpement/project-cnam/src/main/resources/swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Easy eat"
  x-logo:
    url: "/redoc/logo.png"
    backgroundColor: "#FFFFFF"
# host: "localhost:8080"
# basePath: "/"
# schemes:
# - "http"
produces:
- "application/json;charset=utf-8"
...
paths:
...
  /ingredients:
    post:
      tags:
      - "Ingredient"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      operationId: "ingredientCreate"
      summary: "Create an ingredient"
      description: "This endpoint allows to create an ingredient."
      deprecated: false
      parameters:
      - name: "Ingredient"
        required: true
        in: "body"
        description: ""
        schema:
          $ref: "#/definitions/Ingredient"
      responses:
        201:
          description: "Success"
          schema:
            $ref: "#/definitions/Ingredient"
          headers:
            Location:
              description: "ingredient/{ingredientId}"
              type: "string"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/Error"
        412:
          description: "Precondition Failed\n\nList of supported error codes:\n- 65:\
            \ Precondition failed"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/Error"
    get:
      tags:
      - "Ingredient"
      produces:
      - "application/json;charset=utf-8"
      operationId: "ingredientFind"
      summary: "Find an ingredient"
      description: "This endpoint allows to get a list of ingredient."
      deprecated: false
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Ingredient"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/Error"
        412:
          description: "Precondition Failed\n\nList of supported error codes:\n- 65:\
            \ Precondition failed"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/Error"
  /ingredients/{ingredientId}:
    get:
      tags:
      - "Ingredient"
      produces:
      - "application/json;charset=utf-8"
      operationId: "ingredientGet"
      summary: "Get an ingredient."
      description: "This endpoint allows to get an ingredient."
      deprecated: false
      parameters:
      - name: "ingredientId"
        in: "path"
        required: true
        type: "string"
        description: "ingredient identifier."
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Ingredient"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/Error"
        412:
          description: "Precondition Failed\n\nList of supported error codes:\n- 65:\
            \ Precondition failed"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/Error"
    put:
      tags:
      - "Ingredient"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      operationId: "ingredientUpdate"
      summary: "Update an ingredient."
      description: "This endpoint allows to update an ingredient."
      deprecated: false
      parameters:
      - name: "ingredientId"
        in: "path"
        required: true
        type: "string"
        description: "ingredient identifier."
      - name: "Ingredient"
        required: true
        in: "body"
        description: ""
        schema:
          $ref: "#/definitions/Ingredient"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Ingredient"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/Error"
        412:
          description: "Precondition Failed\n\nList of supported error codes:\n- 65:\
            \ Precondition failed"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/Error"
    patch:
      tags:
      - "Ingredient"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      operationId: "ingredientPatch"
      summary: "Update partially an ingredient."
      description: "This endpoint allows to partially update an ingredient."
      deprecated: false
      parameters:
      - name: "ingredientId"
        in: "path"
        required: true
        type: "string"
        description: "ingredient identifier."
      - name: "Ingredient"
        required: true
        in: "body"
        description: ""
        schema:
          $ref: "#/definitions/Ingredient"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Ingredient"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/Error"
        412:
          description: "Precondition Failed\n\nList of supported error codes:\n- 65:\
            \ Precondition failed"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/Error"
    delete:
      tags:
      - "Ingredient"
      operationId: "ingredientDelete"
      summary: "Delete an ingredient."
      description: "This endpoint allows to delete an ingredient."
      deprecated: false
      parameters:
      - name: "ingredientId"
        in: "path"
        required: true
        type: "string"
        description: "ingredient identifier."
      responses:
        204:
          description: "Success"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/Error"
        412:
          description: "Precondition Failed\n\nList of supported error codes:\n- 65:\
            \ Precondition failed"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/Error"
...
  /recipes:
    post:
      tags:
      - "Recipe"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      operationId: "recipeCreate"
      summary: "Create a recipe"
      description: "This endpoint allows to create a recipe."
      deprecated: false
      parameters:
      - name: "Recipe"
        required: true
        in: "body"
        description: ""
        schema:
          $ref: "#/definitions/Recipe"
      responses:
        201:
          description: "Success"
          schema:
            $ref: "#/definitions/Recipe"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/Error"
        412:
          description: "Precondition Failed\n\nList of supported error codes:\n- 65:\
            \ Precondition failed"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/Error"
    get:
      tags:
      - "Recipe"
      produces:
      - "application/json;charset=utf-8"
      operationId: "recipeFind"
      summary: "Find a recipe"
      description: "This endpoint allows to get a list of recipe."
      deprecated: false
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Recipe"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/Error"
        412:
          description: "Precondition Failed\n\nList of supported error codes:\n- 65:\
            \ Precondition failed"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  MeasureUnit:
    description: "This enumeration list any unit of measure for an ingredient."
    type: "string"
    enum:
    - "gram"
    - "liter"
    - "piece"
  Category:
    description: "This enumeration represents the different categories of an ingredient."
    type: "string"
    enum:
    - "butchery"
    - "poultry"
    - "game"
    - "fish"
    - "dairy"
    - "vegetable"
    - "fresh_frozen"
    - "grocery"
    - "cellar"
...
  Error:
    description: ""
    required:
    - "code"
    - "message"
    type: "object"
    properties:
      code:
        description: ""
        type: "integer"
        format: "int32"
      message:
        description: ""
        type: "string"
  Ingredient:
    description: "This class represents a food ingredient."
    required:
    - "idIngredient"
    - "name"
    - "measureUnit"
    - "category"
    type: "object"
    properties:
      idIngredient:
        description: "Ingredient identifier."
        type: "string"
      name:
        description: "Ingredient name."
        type: "string"
      measureUnit:
        $ref: "#/definitions/MeasureUnit"
      category:
        $ref: "#/definitions/Category"
...
  Recipe:
    description: "This class represents a recipe from a user."
    required:
    - "idRecipe"
    - "name"
    - "nbPeople"
    - "ingredientList"
    type: "object"
    properties:
      idRecipe:
        description: "Recipient identifier."
        type: "string"
      name:
        description: "Recipient name."
        type: "string"
      nbPeople:
        description: "Number of people expected for the recipe."
        type: "integer"
      ingredientList:
        type: "array"
        items:
          $ref: "#/definitions/Ingredient"
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /cnam-l11/unKnown/1.0.0
schemes:
 - https
 - http